Created Ability Score Boxes and UI Grid
I added a 10 x 10 grid that we can use for ui layout. If needed it can be expanded for more precise measurements of sections. When we are done we can remove the colored boxes, but they are a good way to help visualize it for now. I also added a very basic stats block as an example of how to utilize the grid. It is currently filled with junk data and we can automate stuff later, but the framework is in. We can also make it prettier at a later time.